backend/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
PyMySQL==1.1.1
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
cryptography==43.0.0
